requirements.txt
python-telegram-bot==13.3
requests
pycryptodome